src/e/edit/ETextArea.java, src/e/edit/ETextWindow.java: only set a font based on the context when filling with content, not on every call to ETextArea.setText, because the latter means that we undo the user's deliberate choice of the non-default font when (say) Find/Replace makes its changes to the text using setText.
src/e/edit/FindAndReplaceAction.java: wrap the call to setText in a
CompoundEdit so that C-Z afterwards hides the fact that JTextComponent
implements setText as a removal of all content followed by an insertion
of new content. We leave the caret in the wrong position, still, but
that's a general problem with our undo/redo functionality (though this
isn't because the underlying UndoableEvents don't update the caret or
expose enough information for us to do so; it's because a bulk replace
by its very nature obliterates knowledge of where the caret was).