src/e/util/Bag.java: why isn't Bag a standard collection class? I haven't seen a bag since university formal specification lectures, but I have written code that would have been better for the use of one...
src/e/edit/Indenter.java: guessing indentation is an example of code
crying out for a Bag rather than a Set. The previous heuristic of choosing
the shortest indentation string was stupid, but easy. What we're really
interested is "which is the most frequent indentation string?", and a
Bag helps us answer that. We have to work a bit harder to get all the
relevant indentation strings than we were doing. And we should accept
tabs as well as spaces!