src/e/edit/ETextArea.java: removes dead code and unused imports.
src/e/gui/EPopupMenu.java: refuses to end a menu with a separator. It's
easier to remove them here than it is to write all the menu-constructing
code in such a way that it can't arrive at such a situation. (The idiom
being that you add your bunch of items followed by a separator, and pass
the partly-constructed menu on to the next interested party.)