src/e/edit/ECaret.java, src/e/edit/ETextArea.java, src/e/edit/MatchingBracketHighlighter.java: moves the code to highlight matching brackets out into its own class. Removes unnecessary JTextComponent parameters from ECaret methods (ECaret, like Caret, knows what JTextComponent it refers to anyway).
Martin suggested we should highlight both brackets in a matching pair,
rather than just the other one (as we did previously). He reckons it will
be useful, whereas my only temptation before had been that I disliked the
asymmetry. It certainly does look better now, and it seemed too ironic to
highlight balanced brackets with an unbalanced visual effect.

Changes the matching bracket highlight color to a pastel green; we haven't
used green yet, and I found the purple a little too close to the selection
color. Now we highlight both, I'm not convinced the 'far' bracket
shouldn't be highlighted in a different color. Maybe there is an inherent
asymmetry in the information we're trying to present?

The other thing that now strikes me is that there's no visual feedback
when you type a matching closing bracket, because we've always considered
the active side of a bracket to be the inside (which gives much less
weird double-click semantics than you see elsewhere).