src/e/gui/KeyboardFocusMonitor.java: provides a simple and convenient interface to observing the current KeyboardFocusManager's focusOwner property.
src/e/gui/ETitleBar.java, src/e/edit/TagsPanel.java: instead of being
prodded by EWindow and ETextWindow respectively, take control of our
own destiny by observing the focus owner ourselves. An interesting
difference to note is that we now only have one KeyboardFocusMonitor
dealing with the focus concerns of all title bars (or single tags panel),
rather than the old situation where every window was concerned with a
title bar and the tags panel. Many-to-one and many-to-one have become
one-to-many and one-to-one, and there's been a related decrease in
the number of live objects. This wasn't the reason for this change,
though: the reason was that no class should implement a Listener
interface unless doing so is the class' primary purpose. Changing the
color of title bars and keeping the tags panel up-to-date should have
been the concerns of title bars and the tags panel.

src/e/edit/EErrorsWindow.java, src/e/edit/ETelnetWindow.java,
src/e/edit/ETextWindow.java, src/e/gui/EWindow.java: removes the
old FocusListener stuff, which is no longer needed.