src/e/util/Log.java: there was a suggestion in the TODO file that we should have a thread outputting regular timestamps to the log, but I couldn't think why we'd want such a thing. Martin suggested it was so that if we see an unexplained exception in there, we at least have some idea of when it was thrown.
We can do better than that, though, and I've been meaning to investigate
it for some time. I was looking through the JDK source once when I came
across EventDispatchThread which is the code that's usually responsible
for the "Exception occurred during event dispatching:" stack traces. It
turns out that there's also a secret system property that you can set to
offer your own handler for uncaught exceptions in the event dispatcher
thread.

They say they'll take it away in future, but suggest that they'll offer
a more public alternative, so we can switch to that if/when the time
comes.

TODO: remove the job we won't be doing.