Changed the input stream reader to read chunks of bytes rather than reading one byte at a time.
Added a new TelnetAction, a simple interface for performing some action on
a TelnetListener.  Everything which is passed on to a TelnetListener is now
implemented as a TelnetAction (except for the processActions(TelnetAction[]) method).

Added TelnetListener.processActions(TelnetAction[]) so all actions created from reading
a buffer-full of input can be processed together.  An array of actions is passed to the
TelnetListener in the AWT dispatch thread.  This allows, in future, the TelnetListener to
store up all the redraw requests and things until after all bunched actions have been
performed, so reducing the redraws the user never sees, and hopefully resulting in
quite a significant speed increase.

It's already considerably faster than before, just because of the bunching of actions.