src/e/edit/AutoCompleteAction.java: adds simple auto-completion. Start typing an identifier, hit alt-esc, and Edit will pop up a list of identifiers you might have meant.
src/e/edit/ETextArea.java: provides a new method to get the text of the
word leading up to the caret, so we know what we're trying to complete.

src/e/edit/EditMenuBar.java: adds the "Complete" option to the "Edit"
menu, with its funky alt-esc keyboard equivalent.

src/e/edit/JavaResearcher.java, src/e/edit/Advisor.java: change the
JavaResearch to offer a single shared instance, for the convenience of
the current AutoCompleteAction. Work through the summary JavaDoc,
extracting the unique identifiers, so we can easily return a list of
identifiers starting with a given prefix when asked.