Split Style out of StyledText.  StyledText now has a Style instead.  StyledText still contains all the conversion magic for VT100-colours to java.awt.Color colours, which should probably be moved elsewhere, but that can be left for later.
Created a Highlight, which defines a start and end Location and a Style (which has zero or more
attributes set).  The Highlighter knows how to convert an array of StyledText objects into another
array with its overloaded styles emblazoned upon them.

Modified JTextBuffer with a simple hack to test the highlighting.  Basic support is there for adding
and removing Highlighters, but that's not used yet.