Highlights may now have a Cursor set, to change the mouse pointer when it hovers over them. If the Cursor is null (ie it hasn't been explicitly set) the default cursor is used.
The Highlighter now has an extra method to optionally perform some action when one of
its highlights is clicked.

Added a HyperlinkHighlighter to find and highlight hyperlinks in fg=blue, underline=true.
This highlighter changes the cursor to the hand thingy, and for now has an implementation
of the highlight clicked method which prints out the URL clicked on.

Fixed a little bug with the CharSequence generation - it now actually works.