Refactored the escape sequence handling completely.
Escape sequences are now sent to an EscapeParser, until this
parser returns true when asked if the sequence is complete.
Within the EscapeParser a SequenceRecogniser is used to recognise
when a sequence is complete (the recogniser implementation used is
decided based upon the first character in the sequence).  This
allows us to parse the OSC sequences, previously given their own
special hack within TelnetControl, in the same place.

Once parsing is complete, the EscapeParser is asked for a
TelnetAction, which is generated by the SequenceRecogniser.  The
actions are found within the new 'escape' package, which contains
one action per type of escape sequence, which helps to cut up the
source into nicer-sized files.

I've also added handling of the cursor save/restore functionality.

There's the possibility that OSC sequences (now referred to as
XTerm sequences, since that's what rxvt calls them) might be
mishandled in some cases - it's not been thoroughly tested.

However, all in all we should have some much more robust and
correct escape sequence handling.