Put in tab position setting/clearing, and the code to use set tab stops rather than the default 'every 8 characters' positions.
Tabs are also stored in the model, so they're not converted to
spaces and lost forever.

Selections of text copied to the clipboard now contain tabs where
they should.

The only remaining tabby thing which needs to be done is to get
the mouse selection to highlight entire tabs, rather than the
individual spaces which make up their display.