Maybe not everyone would like it, but it's a way to start discussion :-)
Terminator now always has a tabbed pane, even if it only has one
tab.  We might change this later, but for now it's easier this
way.

New hotkeys are available.  Alt-n opens a new shell tab, alt-e
opens a little dialog prompting for a command to execute, and then
executes it in a new tab.  Alt-left/right arrow switch between the
tabs.