Hyperlink recognising and following is now decided by properties files.
The user is encouraged to place a file called '.terminator' in his
home directory, which is of a format described by the example
files in 'doc/props-examples/'.  The KDE one works on my box, and
the MacOSX one *should* work, but I don't have any way to test it.

I've added 'mailto:' aswell - Elliott, please check that 'open
foo@bar.com' is going to do the right thing.  Ta.

The format of the properties files is a little unconventional, but
works nicely for setting up hyperlinkers.  We can change it later
if needs be.