The xterm-color derived terminfo wasn't flying; I've instead copied and modified the terminfo file from rxvt (since I've been using its source for implementation ideas).
I had to fix one bizarre bug, which is that rxvt's terminfo
suggests an extra character on the end of the escape sequence to
turn off all font style attributes, whereas for us these extra
chars appear as little boxes.  And, of course, I've switched off
automatic margins.

We now call our terminfo file 'terminator'; xterm-nam is a bad
name, because it's no longer derived from xterm's terminfo file;
we also may want to modify our terminfo file in the future to fix
bugs and things, so let's give it a name which indicates how we're
going to proceed.