src/e/edit/EColumn.java: getNonErrorsWindowCount is more obviously implemented by saying what it is rather than what it isn't, now there's a way of getting the collection of text windows from the workspace.
Fix an exception in some previously unreached code in moveTo which had a
+-1 error in a comparison between an index and a container count.  Make
such errors less likely by using Components rather than indices where
possible.

src/e/edit/EColumn.java, src/e/edit/ETitleBar.java: We couldn't find any
reason not to implement the FIXME that said that EColumn.moveBy should be
called moveTo.

src/e/edit/EColumn.java, src/e/edit/EWindow.java: move the EWindow-specific
part of EColumn.removeComponent into, er, EWindow.  This allows
Workspace.takeWindow to remove the window from one Workspace and add it to
another without breaking it in the process by calling some of the window
closing code.

src/e/edit/Edit.java: try to move files to other workspaces before failing
to removeCurrentWorkspace because it's non-empty.  This requires the
ability to find the second-best workspace for the files on the current
workspace.  Elliott suggests that a more elegant solution would be to
remember all of the files in the current workspace, delete it and then
reassign those files to each one's best workspace.

When adding a workspace, see if any files from other workspaces would be
better placed on the new workspace and, if so, steal them.  This required
some black "revalidation" magic to ensure that the window placement on the
new workspace was non-overlapping.

src/e/edit/Workspace.java: move files to the best workspace for use when a
workspace is removed or added.  This was facilitated by having a
getTextWindows method which, when factored into other, existing code,
removed the need for a few, nasty-looking (to C++ eyes) instanceof checks.