terminator: obviate the need for terminator-dist-script by adding the jar file to the classpath...
terminator-dist-script: ... rendering this file obsolete.

Makefile: rewrite Phil's jar/DIST rule as a rule called "bindist" in
salma-hayek (Elliott suggests that this (or something reasonably
homophonic) is a standard gnu name).  The resulting file is now called
$(PROJECT_NAME)-bindist.tgz.  The -bindist suffix is to avoid confusion
with Elliott's "dist" rule.  If we renamed the project to be terminator
instead of TerminatorN, then the prefix would be fixed automatically.
The motivating reason for switching to .tgz from .zip was just my
familiarity with the command.  Other reasons would be that zip isn't so
commonly installed on Unix boxes and its dubious intellectual property.
We could build this all the time - it's free (for me, anyhow).