Makefile: project.rpf: terminator: everyone was confused by the "N" in TerminatorN.  Phil explains:
We can probably drop it.  Originally the idea was that instead of the number
the films use - 1, 2, 3 etc - this is called Terminator N, for any value of
N.  The significance of this is that you're able to open N terminal sessions
within the same program.

I've gone off the idea though, and the only reason the 'n' is still there is
pure laziness.