Implemented [71G escape code so some guy in Brazil can set the caret horizontal location.
Implemented a way of changing colours - right-click on a coloured
thing (or on the background), and you get a menu whose last item
says 'Change Colour <foo>...', where <foo> is the name of the
colour.  Click on that and you get a colour chooser dialog.  Once
you've said OK, it saves the modified choice to your .terminator
file, and repaints the screen so your colours update immediately.