src/terminator/view/JTerminalPane.java: tells the HyperlinkHighlighter what the current window title is...
src/terminator/view/highlight/HyperlinkHighlighter.java: ...so that if your
prompt sets the window title to the current directory, you have awesome
linking capabilities. Want to click on filenames in 'ls' output? Not a problem.
Want to click on addresses in 'grep' output? Not a problem. This is so cool!