Makefile: Elliott asks:
* presumably there's some easy way to only assign in java.make if a
value hasn't already been assigned.

* an explanation of why variables are rules should go on opposite sides
of the include might be interesting.

So I've explained both of those.

In yesterday's log Elliott said:

> publish terminator to locke

Without any justification for the differentiation between terminator and everything else.  Why make work?  Let's put it in the same place as everything else subject, at least, to there being a reason for doing something else.  Phil doesn't need to be able to login to jessies.org to download the distribution and Elliott's presumably planning on updating it with cron, so Phil doesn't need to be able to write to it either (except via the medium of cvs, which he's happily doing).