java.make: Implemented my understanding of what Elliott meant by this:
> which reminds me; the links to tarballs you've got there just link to
> "edit.tgz" or "scm.tgz"; i was getting used to the date in the filename as
a
> good way of checking whether i was up-to-date or not.

i agree, but i couldn't think of a way to square that with needing an
explicit link in the web page. thinking about it, if we were to make
java.make even more gnarly, we could have a directory
public_html/software/$(PROJECT_NAME)/nightly-builds/ that gets the dated
ones, and a symbolic link to the latest. that target is already getting a
bit complicated. martin?