java.make: Elliott supplied the skeleton of cvs and svn commands for getting a list of all the revision controlled files.  The cvs output is a pain to parse and svn insists on including directory names.  Neither of them work outside the repository.  So sfiles is another thing that bk does better.
Most of the changed lines are tidying changes that looked like they were going to be necessary but actually weren't:

Move the variables before the rules and comment the line between.

Add an echo.<variable name> rule for easy debugging of delayed-evaluation assignments.

Remove the .tar.gz file before recreating it in the make dist rule to avoid gzip asking the user - overwriting is always what's wanted when testing the rule several times on the same day.