Makefile: Move the rules to build pty into java.make.  Elliott's scared that cut-and-paste will cause us to drown in clueless makefiles.  The concern is valid because that's happened at BlueArc.  I think the solution is to ensure that there's no "cleverness" in the makefiles - that all the cleverness is in java.make.
I suggested the idea of SUBDIRS and Elliott said:

i thought of that, but then decided i think i prefer something a bit more anal. we have src/ for Java source, www/ for HTML, and i think it was a mistake not to have the scripts in bin/. i think we should have an (optional) jni/ for JNI stuff (which we aren't using anywhere yet) and an (optional) native/ for non-JNI platform-specific stuff.

i wouldn't be surprised if there were better names for the directories, but i like the idea. "you just put your stuff in the right place, and we'll do everything automatically".

I agree with the idea and I think the names are OK.  The only reason I haven't done anything about it (the make rules changes would be easy) is because of the poor support in CVS for renaming the pty directory.