src/e/edit/UnmatchedBracketHighlighter.java: Tidy up Elliott's optimisation in order to have an excuse to log the reason he did it:
According to top, we're taking 10% more CPU than before unmatched
bracket recognition. There's only so many 10%s we can chew up.

It might not seem like much, and it might seem that we're still "fast
enough" but disable it and typing is visibly more responsive. Just
because you're not consciously aware of the slowdown doesn't mean it
isn't doing psychological harm.

I didn't measure the *effect* of that change. I can't even remember whether the 10% includes it or not. And I don't know what top means by 10% on a dual-cpu machine.

What I found most interesting was that you could tell the difference if you tried them back to back. That you can be around 25% (typing fast, but not holding down a key) and be slow. Obvious, and at the same time not. Cpu utilization is easier to measure than interactive fell, though.

Our own event dispatch thread that keeps statistics?