src/e/edit/UnmatchedBracketHighlighter.java: Braces-within-parentheses need to be allowed because of this Java idiom:
addXListener(new XListener() {
  // stuff
});

Stacking the bracket type as well as the indentation solves this and removes the divergent duplication between the handling of parentheses and braces.  Someone who knows Java might like to don the rubber gloves and tidy-up the C++ BracketSite class I've added.

I can't help feeling that this should have broken the recovery change but I've failed to demonstrate how.  Which is to say that I've looked at LinuxInstall.txt (renamed to LinuxInstall.cpp) and I've failed to edit it into a state where the highlighting looks unreasonable.