src/e/util/SpellingChecker.java: Ed thinks all suggestions (rather than just the first) should be considered when checking for a near-miss. I can't remember any reason why we only consider the first suggestion, and he offers an explanation of why this might not be what you want:
In my personal dictionary I've got "Phylosopher", which is itself a near-miss
for a proper word, but in these parts it's considered to be a real word.
ispell, however, seems to prefer the words from the main dictionary when it
offers suggestions. So, when I spell it with a small 'p', I'm offered

[ed@lexone Projects]$ ispell -a
@(#) International Ispell Version 3.2.06 08/01/01
phylosopher
& phylosopher 2 0: philosopher, Phylosopher

and SpellingChecker rejects it as being equal except for case. I don't think
it should, hence a patch for your consideration.