src/e/edit/Edit.java: Ed says:
Filename links in the errors window aren't opened correctly when they're
absolute paths containing elements like a/b/../c/d. openFile should convert
the filename to the canonical path, not an absolute one.

you can test it by getting some file links in the error window, making them
absolute, and putting .. elements in, and see that different windows open on
the same file, e.g

1) mangle ~/Projects/edit/src/e/edit/Edit.java so it doesn't compile
2) click the link, Edit.java opens
3) edit the link text to ~/Projects/edit/src/e/../e/edit/Edit.java
4) click on it, a new window "~Projects/edit/src/e/../e/Edit.java" opens

-ed

The comment above this line in the code confirms that the intention of the call to getAbsolutePath was, indeed, the removal of /../ sections of the path.  The javadoc confirms that it's getCanonicalPath that does that, not getAbsolutePath.