src/e/edit/Edit.java: Ed says:
in fact the last line of that try block is all that's necessary. if we're going to canonicalize the path anyway, why bother checking if it's a symbolic link first?

I said:

History doesn't go back far enough to record the reason.  I've a nasty feeling that I've tried this before.

...

I think we've figured it out.  The reason it didn't used to getCanonicalPath was that getCanonicalPath would change /home/martind/kipper/Makefile into /u/u58/martind/kipper/Makefile, which would put the file on the default (rather than the desired) workspace.  We had no choice but to follow the link for symlinks, or we'd end up writing new files into our include directory and then renaming them over the top of the symlink to the correct location.  So, if we tried to edit a symlink, we'd end up on the default workspace - which wasn't very satisfactory.  That all got fixed when this code went in:

    /**
     * Substituting ~/ for /home/userName isn't the only friendly transformation
     * that we can usefully and easily make.  If the user has configured us
     * with workspace paths which include symbolic links, then it's a reasonable
     * presumption that the user would like us to consistently use the form of
     * the path which they supplied rather than the OS-canonical form (supplied
     * by File.getCanonicalPath).
     * A symbolic link within such a workspace would have both symbolic links
     * canonicalized in openFile and so would end up with a path prefixed by
     * the desired Workspace's canonical name rather than its friendly name.
     * This would send it to the wrong Workspace.
     * FIXME: Consistency probably demands that this be called from
     * FileUtilities.getUserFriendlyName rather than Edit.openFile.
     */
    public static String normalizeWorkspacePrefix(String filename) {

We did note, though, that the getCanonicalPath code is happening before any parsing of "~/".  I haven't managed to come up with a test that demonstrates a problem with that.  I don't think that affects whether the isSymbolicLink check should come out.  I think it should come out but Elliott may be having some ..-related problems that he might like to share first.

And Elliott said:

if i'm building in libs/vlsi/fsm and i've buggered one of the Perl scripts, Perl reports the error using a filename containing "/../", and -- if you click on it -- Edit will open that exact name, rather than taking you to the copy you've already got open without the dots. this was fixed by Ed's change.

syntax error at ~/Projects/nelson/libs/vlsi/nim2/../make-release-constants.pl line 5, near "my "
Execution of ~/Projects/nelson/libs/vlsi/nim2/../make-release-constants.pl aborted due to compilation errors.

In it goes.