src/e/edit/TagsPanel.java: remove a lot of code that belongs in TagsUpdater now.
src/e/edit/TagsUpdater.java: gain code from TagsUpdater, and change it to
move away from SwingWorker (which always seems over-complicated and rather
awkward to use), and to re-use the same temporary file for the tags so that
the MD5 digests are comparable: it turns out that Exuberant ctags writes the
name of its output file into the output file! [The fix to my digest code,
and the move away from SwingWorker came from Ed; I just moved the classes
out of TagsPanel, and made them simpler in the process. The good news is he's
bought a PowerBook, so we soon won't have to worry about regressions caused
by him being out of date and us not reading patches carefully enough.]