src/e/util/StringHistory.java: moved from the Edit repository. Added the ability to cope without a backing file. Added a method to clear the history.
src/e/gui/InteractiveBorder.java: factors out and generalizes the code for
the cancel button so we can use it for the recent searches menu too.

src/e/gui/SearchField.java: adds a recent searches menu. How does it know when
to add something? For the notification-on-Enter case, it's easy. But it's not
obvious what to do when we're sending notification on each keystroke. Until I
have some experience to go on, items are only added to the history when you
hit Enter, regardless.