src/e/edit/ETextWindow.java: Add a comment to allow me to paste in the "why" behind my claims which Elliott reported in the previous check-in.
File.renameTo would break symlinks if we ever edited via the original path rather than following the symlinks.  This would happen on the current code if, for example, we were running over cifs.  In the email thread, I described this as "a remote possibility" but, thinking about it, Elliott uses cifs between his Macs.  I don't know if the unix extensions allow client-side symlink following or whether it's done server-side as in the case of Microsoft clients accessing Samba, billbox or titan.

Then I asked what the motivation was for breaking symlinks.  Elliott came up with a couple of reasons:

o  ETextWindow.copyFile had, for years, had a comment saying "FIXME: replace with File.renameTo"

o  making a copy of the file is more similar to writing the new version of the file than renaming is.  If making the copy fails to allocate disk space, that indicates that writing the new version is likely to fail in the middle.  We don't want that to happen as it would leave us with no copy of the user's precious input in stable storage.

A permissions problem could also stop us from writing.  If you can write the file but not the directory, neither renaming nor copying will work.  Perhaps that's the reason that we've seen MS Word (at Fred Hutchinson) save its temporary files somewhere else (ironically, somewhere that it doesn't have permissions).

Elliott adds:

> one alternative i had was to save to a temporary file (and let the system 
> decide where that goes), but that's not very good if you're trying to 
> check out the file system you'll be writing to.
>
> but you can't really do the right thing (guarantee that you either 
> leave a perfect copy of the new data, or don't change anything).

We talked, albeit tangentially, about disk-full and permissions problems but what, perhaps, convinced Elliott was when I brought up fail-stop problems like power outages (or server crashes or network partitions).

I argued that the most important constraint is to insist that there's always at least one complete, consistent copy of the user's input in stable storage.  That's what suggested writing out the new version twice, rather than copying the old version of the file.  I think this is what Word does.  Like vim, it will have an auto-recovery check when it starts up to see if it crashed during writing of the second copy (to the original location).

I didn't bring this up because I didn't want to dissuade anyone from making the change to write the new version twice, but I did use the term "stable storage" carefully.  Meaning that a write or a close to a file system might seem to be committed but there's no sure way of knowing that it's in stable storage.  Reading the file back will likely get the information from cache.  The longer the time since the write, the more likely it has made it to stable storage.

This should reinforce the point Elliott made in the comment he added to ETextWindow:

// Ideally, we should do this from a timer and always have a recent backup around.

It suggests that we should try to space-out the writes to the two files.  So perhaps the auto-saving writes only go to the second copy and the final write only goes to the original location.