src/e/edit/ShellCommand.java: When the --login was commented out (pre-CVS-ChangeLog), no reason was given.  The need for --login - to set ANT_HOME - had just been removed.  Elliott thinks that it was probably just tidying up - removing unnecessary things.
Now we want to introduce a BlueArc-specific hack in misc/startup_scripts/profile (or similar) to call switchbuild when in a BitKeeper repository so we can build with Edit out of the box on anyone's installation.  As Elliott points out, a hack in the makerules would be insufficient because all of the makefiles assume $(SOFTWARE) in order to find the makerules.