edit.properties-sample: JohnW was seeing .d2 files from -eabi build directories.  Replace the fuzzy BlueArc build directory regular expression with a more precise one that I probably stole from elsewhere.
I seem (~martind/.e.edit.Edit/edit.properties) to be ignoring .orig files too, which sounds sensible.  Merge that in too.

edit.properties-default might be a better name than -sample for a file that's automatically installed.  Better still would be if we could update all BlueArc installations with the same list of extensions.  We like jackboots!