src/terminator/view/highlight/HyperlinkHighlighter.java: major (well commented) changes to the regular expression used to recognize filenames. The motivation was fixing the denial of service caused by anything like:
  perl -e 'print "x" x 100_000 . "\n"'

The fix consists of two parts. The first is the new WorkDeferrer class, that
uses a Timer to put off actually doing any work until we stop being asked to
do work. There's an annoying visible delay in highlighting the links, but we
can live with that for now. The other part to the fix is an optimized version
of the old regular expression that eliminates the expensive backtracking. (The
new regular expression also fixes a few problems with the old one's false
positives and negatives.) Markus Laker came up with the optimized expression.