src/terminator/terminal/TerminalControl.java: simplify the course of bytes read from our pseudo-terminal through to being run as TerminalAction instances on the event dispatch thread, primarily by using more conventional idioms. There was no need to cope with processBuffer leaving unprocessed characters in the buffer, because it never does. This realization simplified both caller and callee.
While I'm here, I've made the input buffer bigger because I can see (with
some extra debugging output) that we can easily read more than 4KiB
at a time. I've only doubled it to 8KiB, mainly because that makes the
debugging output easier to read for the thing I'm testing at the moment. We
should revisit this and experiment to find an optimum. We can certainly afford
a much larger buffer, if it buys us anything.

I've added synchronization to all the methods that use the terminalActions
field, because I don't think I understand the concurrency here, and am
unconvinced that this code is safe. Noteworthy is the fact that removing
the flushLineBuffer from processBuffer fixes the bug I'm seeing where the
first time I run

  perl -e 'print "x" x 100_000 . "\n"'

in a new window, I don't actually get 100000 characters (though I do on
later runs).

I've added sendString as preparation for more efficient pasting; I'm not ready
to check that in until I've demonstrated that it's useful.