src/terminator/view/JTerminalPane.java: Remember when I first used debconf in terminator and complained that it wouldn't let me select any of the options?  I felt like an eedjit when Elliott watched me hammering return to no avail, reached over and pressed space, which worked fine.  JohnW points out that whiptail --yesno text 10 10 allows return or space to be used in xterm.  I've looked at why (using tty to see which pseudo-terminal I'm connected to, then running whiptail and, from another window, doing stty -a < psuedo-terminal) and discovered that the newt or SLang library that's behind whiptail turns off "icrnl" on the pseudo-terminal.  In xterm, this results in the application receiving \r when the Enter key is pressed.  You can demonstrate this by doing:
stty -icrnl
cat

And then pressing Enter.  In xterm, you see ^M.  In terminator, you see two newlines.

The newt library checks for either of \r or space to accept the currently selected "button" but not \n.

I think we should be pumping \r at the terminal and letting it do the translation to \n if icrnl is enabled.  I think the Java is the Right place to do this.

The problem's nothing to do with the unsupported escape sequences (which should affect the codes generated by the keypad (who cares?) and the graphics character set):

2004-12-01T13:20:46.306-0800 Terminator: Unsupported single-character escape "=" (set private mode PrivMode_aplKP (application keypad).).
2004-12-01T13:20:46.306-0800 Terminator: Unsupported two-character escape "(B".
2004-12-01T13:20:46.307-0800 Terminator: Unsupported two-character escape ")0".
2004-12-01T13:20:57.885-0800 Terminator: Unsupported single-character escape ">" (unset private mode PrivMode_aplKP (application keypad).).

That looks like a fruitful line of investigation.  A lingering minor complaint that I've never mentioned is that, in less, if you're searching with / and editing the regular expression, backspace inserts a ^H rather than backspacing.  If you run rxvt and do Ctrl-V backspace, you see ^?.  In terminator, you see ^H.  In xterm, you see ^H too but xterm has the same problem in less that terminator does.  Converting the backspace to serial terminal-style ^? fixes less without breaking vim or bash's line editor (or the billbox's), as far as I've seen.

Elliott said:

> now you're persuading me. i've actually noticed this one (despite having set PAGER=cat). 
> what i'd do at this point is have a look at the xterm (or, more likely, rxvt) source to see if i could find anything equivalent. the ^? sounds plausibly Unix, but '\r' seems a little odd. though not completely unbelievable. i mean, that's the kind of thing that was probably decided back when they were still typing # to "erase" a character and @ to "erase" a line. 

That code turns out to be:

                case XK_KP_Enter:
                /* allow shift to override */
                    if ((r->h->PrivateModes & PrivMode_aplKP) ? !shft : shft) {
                        STRCPY(kbuf, "\033OM");
                    } else {
                        kbuf[0] = '\r';
                        kbuf[1] = '\0';
                    }
                    break;

...

#ifndef NO_BACKSPACE_KEY
                case XK_BackSpace:
                    if (r->h->PrivateModes & PrivMode_HaveBackSpace) {
                        kbuf[0] = ((!!(r->h->PrivateModes & PrivMode_BackSpace)
                                    ^ !!(shft | ctrl)) ? '\b' : '\177');
                        kbuf[1] = '\0';
                    } else
                        STRCPY(kbuf, r->h->key_backspace);
# ifdef MULTICHAR_SET
[snip]
# endif
                    break;
#endif

Shift doesn't seem to have any effect on what gets generated for Enter in rxvt for me.  Shift or Ctrl cause backspace to generate ^H though, as the code might suggest.  I was hoping rxvt might have a good name for this "uncooking" function, but it doesn't:

/*{{{ Convert the keypress event into a string */
/* INTPROTO */
void
rxvt_lookup_key(rxvt_t *r, XKeyEvent *ev)

The XK_KP_ENTER code convinced Elliott.  He added:

> do we know what uses the other keyboard?

To which I replied:

> The PrivMode_aplKP nonsense?  Well, whiptail for one:
> 2004-12-01T13:20:46.306-0800 Terminator: Unsupported single-character escape "=" (set private mode PrivMode_aplKP (application keypad).). 
> Although, like I say, I can't seem to get shift to toggle the Enter behavior in rxvt anyway.

And:

> are there many places like this in the source, where we'd have to do something different? if so, i might do that when i get chance, and have some swappable key translator. 

To which I replied:

> Yes, though we seem to be getting clean away with it.  Although I hadn't reported either of these problems (except, sort of, the Enter/whiptail one), I haven't noticed any others.

And, regarding XK_BackSpace:

> looks like there's maybe some mode where you can have an arbitrary string for backspace, or there's some other mode that decides whether it's '\b' or '\x7f'. wonder if anything uses those modes? 

(Gah - I thought I'd committed this on the date of the last email on this topic - 2004-12-03!)