doc/terminfo/terminator.tic: Simon complained [
I just typed "vi memory.cpp" in a terminator window, and in the parent, this appeared.

2005-02-21T14:56:05.690+0000 Terminator: Unsupported single-character escape "=" (set private mode PrivMode_aplKP (application keypad).).
2005-02-21T14:56:08.329+0000 Terminator: Unsupported single-character escape ">" (unset private mode PrivMode_aplKP (application keypad).).

]

Elliott noted that this has always been a known issue [

yes. those escape sequences are unsupported in Terminator. they aren't important in any application i know of, but i don't want to take the warnings out in case one day they're the reason we're having trouble with something.

ideally, we'd be able to say in our termcap entry that we don't support this. (can we?)

]

Martin said "yes we can" [

man 5 terminfo has this barely comprehensible gibberish to say.  The last two sentences are the important ones but they make slightly more sense in context:

       If  the  terminal  has  a keypad that transmits codes when the keys are
       pressed, this information can be given.  Note that it is  not  possible
       to handle terminals where the keypad only works in local (this applies,
       for example, to the unshifted HP 2621 keys).  If the keypad can be  set
       to transmit or not transmit, give these codes as smkx and rmkx.  Other-
       wise the keypad is assumed to always transmit.

That suggested <this patch>.

That passed the spot test for me - I don't get the warnings and I can still move the cursor in vim.  It's not been long enough for a regression test.

]

Elliott said "sheep eat" [

i'd be happy with this patch, though. if you want to keep trying it for a bit first, that's fine too.

]